Restructure Mock API
